third_party/WebKit/LayoutTests/http/tests/serviceworker/ServiceWorkerGlobalScope/registration-attribute.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: registration</title>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src='../resources/test-helpers.js'></script>
<script>

promise_test(function(t) {
    var script = 'resources/registration-attribute-worker.js';
    var scope = 'resources/scope/registration-attribute';

    return service_worker_unregister_and_register(t, script, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(scope); })
      .then(function(frame) {
          assert_equals(
              frame.contentDocument.body.textContent,
              'PASS',
              'Service Worker should respond to fetch');
          frame.remove();
          return service_worker_unregister_and_done(t, scope);
        });
  }, 'Verify registration attribute on ServiceWorkerGlobalScope');

</script>
