third_party/WebKit/LayoutTests/platform/linux/virtual/mojo-loading/http/tests/serviceworker/ServiceWorkerGlobalScope/registration-attribute-expected.txt

This is a testharness.js-based test.
FAIL Verify registration attributes on ServiceWorkerGlobalScope assert_equals: Service Worker should respond to fetch expected "updatefound,install,statechange(installed),statechange(activating),activate,statechange(activated),fetch" but got "\nNot Found\nThe requested URL /serviceworker/ServiceWorkerGlobalScope/resources/scope/registration-attribute was not found on this server.\n\nApache Server at 127.0.0.1 Port 80\n\n"
FAIL Verify registration attributes on ServiceWorkerGlobalScope of the newer worker promise_test: Unhandled rejection with value: object "Error: wait_for_state must be passed a ServiceWorker"
Harness: the test ran to completion.
